gerrymandering-mcmc-master1/requirements.txt
astroid==2.3.3
cycler==0.10.0
decorator==4.4.1
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
networkx==2.4
numpy==1.17.4
pathlib==1.0.1
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
mpi4